mcp-server/requirements.txt
boto3>=1.34.0
mcp>=1.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-mcp-adapters>=0.1.0
